requirements.txt
onepassword-sdk
onepasswordconnectsdk
typing-extensions
python-dotenv